Adding Quests (#3)
* Add README with examples of daily quests for Credify plugin

This README contains a variety of daily quest examples tailored to different playstyles and engagement levels. It outlines gameplay, engagement, progression, and unique quests, along with implementation tips for better quest management and player engagement.

* Add quest system and various enhancements

Introduce a new quest system for tracking and rewarding player activities. Implemented support for quest-based events across various game commands and activities. Refactored some existing classes and methods to support the new quest logic and event raising capabilities.

* Remove README and add new quest types

Deleted outdated README file and introduced new quest types: Humiliation and Trivia. Adjusted QuestManager and ChatGames to support new objectives.

* Update payout condition for 'Baller' event invocation

Changed the condition to raise the 'Baller' event when the payout is greater than or equal to 10,000 instead of just greater than 10,000. This update ensures that payouts exactly at the threshold will also trigger the event across both Roulette and Blackjack games.

* Update event services, translations, and quest configurations

Improved event service calls by passing required data, updated various translations from "lotto" to "raffle," and refined quest configurations for better clarity and functionality. Additionally, fixed minor issues and removed redundant code for improved readability.

+++ Credify/Chat/Active/Roulette/Models/BetTypes/BaseBet.cs
@@ -2,7 +2,7 @@ namespace Credify.Chat.Active.Roulette.Models.BetTypes;
 
 public abstract class BaseBet(int stake)
 {
-    public int Stake { get; } = stake;
+    public long Stake { get; } = stake;
     public abstract int Payout { get; }
     public abstract bool HasWon(SpinResult spinResult);
 }
